gnusocial-next/workers/media-processing/requirements.txt
redis==5.2.1
pillow==11.1.0
